requirements.txt
fastapi
Pillow
numpy
tensorflow
uvicorn
